fix(cli): typo, etc
