+++ src/cli/validation.ts
@@ -1,7 +1,7 @@
 import { validateMnemonic } from 'bip39';
 
 export const mnemonicValidation = async (input: string) => {
-  if (validateMnemonic(input)) {
+  if (!validateMnemonic(input)) {
     return 'Invalid mnemonic';
   }
   return true;
